+++ tests/real_kernels_test.py
@@ -1,0 +1,9 @@
+from .conftest import template
+
+
+class TestRealKernels:
+    def test_mask_kernel(self):
+        template('real_kernels', 'mask_kernel')
+
+    def test_weighted_sum_kernel(self):
+        template('real_kernels', 'weighted_sum_kernel')
